+++ rentals.txt
@@ -1,2 +1,4 @@
 {"date": "27/04/23", "person_id": "128", "book_title": "La corruptrice"} 
 {"date": "27/04/23", "person_id": "793", "book_title": "Les larmes de la patience"} 
+{"date": "27/04/23", "person_id": "472", "book_title": "Alice au pays des meiveilles"} 
+{"date": "27/04/23", "person_id": "743", "book_title": "Les larmes de la patience"}
